Catch no-fragment pairs
