+++ .plans/2026-07-01-111806-no-fragment-pair.md
@@ -1,0 +1,59 @@
+# Plan: no-fragment pair detector
+
+## Goal
+
+Catch clipped adjacent `No X. No Y.` prose beats such as `No implant. No brain surgery.` without lowering the general `fragment-stacking` threshold.
+
+## Approach
+
+1. Update `src/rules/syntactic-patterns/repetition/fragment-stacking.ts`.
+   - Keep the existing general fragment stack rule at two fragment types plus at least three sentence units.
+   - Add a separate detector for exactly two adjacent complete sentences that both start with `no`.
+   - Require each sentence to be a short fragment, not a full clause.
+   - Reuse the existing `sentenceWords` and finite-verb checks instead of adding a new parser.
+2. Add hit fixtures to `behavior/fixtures/textlint-rules/cases/syntactic-patterns/hits.md`.
+   - `No implant. No brain surgery.`
+   - `No fever. No rash.`
+   - `No shortcuts. No magic.`
+   - `No dashboards. No extra meetings.`
+3. Add no-hit fixtures to `behavior/fixtures/textlint-rules/cases/syntactic-patterns/no-hits.md`.
+   - `No implant was used. No brain surgery was performed.`
+   - `No fever was recorded. No rash appeared.`
+   - `No parking is allowed after midnight. No smoking is allowed indoors.`
+   - `No sugarcoating the numbers, revenue fell again.`
+4. Verify behavior.
+   - Run the current CLI on targeted probes before and after the change.
+   - Run `fixture3 check --suite textlint-rules-cases-syntactic-patterns`.
+   - Run `fixture3 check --feature textlint-rules`.
+   - Run `pnpm run validate`.
+5. Manually review all new findings.
+   - Inspect fixture3 diffs.
+   - Inspect full fixture/corpus output for `fragment-stacking` additions.
+   - Classify each new hit as desired or false positive before approving any golden drift.
+
+## Matching Rules
+
+Report a no-fragment pair when:
+
+- two adjacent sentences both start with `no`;
+- each sentence has 2-5 normalized words;
+- each sentence has no finite verb from the existing `FINITE_VERBS` list;
+- each sentence is classified as a noun fragment by the existing fragment classifier;
+- the pair is not rejected by `hasConcreteInventorySubjects`.
+
+Do not report:
+
+- full factual clauses with verbs;
+- comma-led sentences such as `No sugarcoating the numbers, revenue fell again.`;
+- general fragment stacks that do not use the `No X. No Y.` cadence.
+
+## Files To Modify
+
+- `src/rules/syntactic-patterns/repetition/fragment-stacking.ts`
+- `behavior/fixtures/textlint-rules/cases/syntactic-patterns/hits.md`
+- `behavior/fixtures/textlint-rules/cases/syntactic-patterns/no-hits.md`
+- `behavior/golden/**` only after reviewing fixture3 diffs
+
+## Key Decision
+
+This belongs in `fragment-stacking`, not `negation-reframe`, because the defect is clipped fragment cadence. It is not a semantic replacement frame like `not X, but Y`.
