+++ .worklogs/2026-07-01-112610-no-fragment-pair.md
@@ -1,0 +1,39 @@
+# No-fragment pair detector
+
+## Summary
+
+Added coverage for adjacent clipped `No X. No Y.` prose beats under the existing `fragment-stacking` rule. The detector now catches pairs such as `No implant. No brain surgery.` and `No fever. No rash.` without lowering the broader fragment-stack threshold.
+
+## Decisions Made
+
+- Kept the rule ID as `syntactic-patterns:fragment-stacking` because the defect is fragment cadence, not a negation reframe.
+- Added a targeted two-sentence `no-fragment` path while leaving the general fragment stack requirement unchanged.
+- Split detector logic into `src/rules/syntactic-patterns/repetition/private/fragment-stack-detector.ts` after validation caught the 400-line file limit.
+- Added no-hit guards for full factual clauses such as `No implant was used. No brain surgery was performed.`
+
+## Key Files
+
+- `src/rules/syntactic-patterns/repetition/fragment-stacking.ts`
+- `src/rules/syntactic-patterns/repetition/private/fragment-stack-detector.ts`
+- `behavior/fixtures/textlint-rules/cases/syntactic-patterns/hits.md`
+- `behavior/fixtures/textlint-rules/cases/syntactic-patterns/no-hits.md`
+- `behavior/golden/textlint-rules-cases-syntactic-patterns/approved.normalized.json`
+- `.plans/2026-07-01-111806-no-fragment-pair.md`
+
+## Verification
+
+- Target probe catches `No implant. No brain surgery.` and `No fever. No rash.`
+- Target no-hit probes stay clean for full factual clauses.
+- `fixture3 check --feature textlint-rules` passes.
+- `pnpm run validate` passes.
+
+## Manual Review
+
+Only four new `fragment-stacking` findings appeared, all in syntactic-patterns `hits.md`:
+
+- `No implant. No brain surgery.`
+- `No fever. No rash.`
+- `No shortcuts. No magic.`
+- `No dashboards. No extra meetings.`
+
+No `fragment-stacking` findings appeared in syntactic-patterns `no-hits.md`. All corpus suites matched after approval, so this change did not introduce corpus false positives.
