+++ behavior/fixtures/textlint-rules/cases/syntactic-patterns/hits.md
@@ -1141,3 +1141,11 @@ Honestly, the rollout was rough.
 Here's the thing: nobody read the memo.
 
 Frankly, the review changed nothing.
+
+No implant. No brain surgery.
+
+No fever. No rash.
+
+No shortcuts. No magic.
+
+No dashboards. No extra meetings.
