+++ behavior/golden/textlint-rules-cases-syntactic-patterns/approved.meta.json
@@ -1,26 +1,27 @@
 {
   "suite": "textlint-rules-cases-syntactic-patterns",
   "kind": "approved",
-  "run_id": "2026-06-02T11-05-49.409965Z-04547efb",
-  "run_commit": "8987c00",
+  "run_id": "2026-07-01T10-21-08.644087Z-04547efb",
+  "run_commit": "69d5e2a",
   "working_tree": "dirty",
-  "recorded_at": "2026-06-02T11:05:49.409965Z",
-  "fixture_hash": "sha256:73dc35becd354b02d659558d0ea0fc42bfc8ae0af739cd6f6073af88a1040649",
+  "recorded_at": "2026-07-01T10:21:08.644087Z",
+  "fixture_hash": "sha256:1d98cd4f96cf2dfafabceb2063cc3188d9a554bafaab21fd1bb168ab2eb68d9c",
   "manifest_hash": "sha256:04547efbb13157ee3d8ef18dda767231fd63d3f9cc27c15317cf28c5f0583e7b",
   "tool_version": "0.1.8",
   "output_schema_version": "1",
   "fixtures": [
     {
       "path": "behavior/fixtures/textlint-rules/cases/syntactic-patterns/hits.md",
-      "sha256": "sha256:26b9feaf28eaafcf883200696b7237d1c3530026c076acfbeeeb8e38b1324cf5"
+      "sha256": "sha256:37b72f98e8afcc48a494236ad78e62f72a6b55318aae3c1b2accbe454a6429a4"
     },
     {
       "path": "behavior/fixtures/textlint-rules/cases/syntactic-patterns/no-hits.md",
-      "sha256": "sha256:db16ee9450ad63bdb12478d918b92df8d4b2f7103e6556f2c982b26a50c895e1"
+      "sha256": "sha256:d713f7ae23482b676cefeffbc2728e71732c3a05eee2eea01f25bd22ba390ac6"
     },
     {
       "path": "behavior/fixtures/textlint-rules/cases/syntactic-patterns/significance-density.md",
       "sha256": "sha256:a5489177104e3714ea3558c1d90bdd91573bf719c83112225ae9e9b0e288c6da"
     }
-  ]
+  ],
+  "comment": "Add no-fragment pair coverage"
 }
