+++ behavior/golden/textlint-rules-cases-syntactic-patterns/approved.normalized.json
@@ -16,7 +16,7 @@
             "line": 1
           }
         },
-        "message": "Flesch Reading Ease is 46.44. Keep it at 61 or higher.",
+        "message": "Flesch Reading Ease is 46.95. Keep it at 61 or higher.",
         "range": [
           0,
           1
@@ -39,7 +39,7 @@
             "line": 1
           }
         },
-        "message": "Gunning Fog is 15.62. Keep it at 12 or lower.",
+        "message": "Gunning Fog is 15.42. Keep it at 12 or lower.",
         "range": [
           0,
           1
@@ -13249,6 +13249,98 @@
         "ruleId": "boilerplate-framing",
         "severity": 2,
         "type": "lint"
+      },
+      {
+        "column": 1,
+        "index": 51303,
+        "line": 1145,
+        "loc": {
+          "end": {
+            "column": 30,
+            "line": 1145
+          },
+          "start": {
+            "column": 1,
+            "line": 1145
+          }
+        },
+        "message": "Fragment stack found: No implant. No brain surgery. Rewrite the clipped cadence as normal prose.",
+        "range": [
+          51303,
+          51332
+        ],
+        "ruleId": "fragment-stacking",
+        "severity": 2,
+        "type": "lint"
+      },
+      {
+        "column": 1,
+        "index": 51334,
+        "line": 1147,
+        "loc": {
+          "end": {
+            "column": 19,
+            "line": 1147
+          },
+          "start": {
+            "column": 1,
+            "line": 1147
+          }
+        },
+        "message": "Fragment stack found: No fever. No rash. Rewrite the clipped cadence as normal prose.",
+        "range": [
+          51334,
+          51352
+        ],
+        "ruleId": "fragment-stacking",
+        "severity": 2,
+        "type": "lint"
+      },
+      {
+        "column": 1,
+        "index": 51354,
+        "line": 1149,
+        "loc": {
+          "end": {
+            "column": 24,
+            "line": 1149
+          },
+          "start": {
+            "column": 1,
+            "line": 1149
+          }
+        },
+        "message": "Fragment stack found: No shortcuts. No magic. Rewrite the clipped cadence as normal prose.",
+        "range": [
+          51354,
+          51377
+        ],
+        "ruleId": "fragment-stacking",
+        "severity": 2,
+        "type": "lint"
+      },
+      {
+        "column": 1,
+        "index": 51379,
+        "line": 1151,
+        "loc": {
+          "end": {
+            "column": 34,
+            "line": 1151
+          },
+          "start": {
+            "column": 1,
+            "line": 1151
+          }
+        },
+        "message": "Fragment stack found: No dashboards. No extra meetings. Rewrite the clipped cadence as normal prose.",
+        "range": [
+          51379,
+          51412
+        ],
+        "ruleId": "fragment-stacking",
+        "severity": 2,
+        "type": "lint"
       }
     ]
   },
@@ -13292,7 +13384,7 @@
             "line": 1
           }
         },
-        "message": "Flesch Reading Ease is 17.51. Keep it at 61 or higher.",
+        "message": "Flesch Reading Ease is 25.33. Keep it at 61 or higher.",
         "range": [
           0,
           1
@@ -13315,7 +13407,7 @@
             "line": 1
           }
         },
-        "message": "Gunning Fog is 27.23. Keep it at 12 or lower.",
+        "message": "Gunning Fog is 24.11. Keep it at 12 or lower.",
         "range": [
           0,
           1
